JeuDeCarte/Steam/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
steamworks-py==1.4.0
